fix(contracts): change Position and EntityType schema to int32/uint32 to fit in js number

+++ packages/ri/contracts/src/components/PositionComponent.sol
@@ -3,8 +3,8 @@ pragma solidity >=0.8.0;
 import "solecs/Component.sol";
 
 struct Position {
-  uint256 x;
-  uint256 y;
+  int32 x;
+  int32 y;
 }
 
 uint256 constant ID = uint256(keccak256("ember.component.positionComponent"));
@@ -21,10 +21,10 @@ contract PositionComponent is Component {
     values = new LibTypes.SchemaValue[](2);
 
     keys[0] = "x";
-    values[0] = LibTypes.SchemaValue.INT64;
+    values[0] = LibTypes.SchemaValue.INT32;
 
     keys[1] = "y";
-    values[1] = LibTypes.SchemaValue.INT64;
+    values[1] = LibTypes.SchemaValue.INT32;
   }
 
   function set(uint256 entity, Position calldata value) public {
@@ -32,7 +32,7 @@ contract PositionComponent is Component {
   }
 
   function getValue(uint256 entity) public view returns (Position memory) {
-    (uint256 x, uint256 y) = abi.decode(getRawValue(entity), (uint256, uint256));
+    (int32 x, int32 y) = abi.decode(getRawValue(entity), (int32, int32));
     return Position(x, y);
   }
